# Conflicts:
#	app/src/main/java/com/iGap/G.java
#	app/src/main/java/com/iGap/activities/ActivityChat.java
#	app/src/main/java/com/iGap/activities/ActivityGroupProfile.java
#	app/src/main/java/com/iGap/adapter/AdapterShearedMedia.java
#	app/src/main/java/com/iGap/fragments/FragmentNewGroup.java
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/iGap/response/GroupLeftResponse.java
@@ -29,42 +29,40 @@ public class GroupLeftResponse extends MessageHandler {
     @Override
     public void handler() {
         super.handler();
-        ProtoGroupLeft.GroupLeftResponse.Builder builder =
-                (ProtoGroupLeft.GroupLeftResponse.Builder) message;
+        final ProtoGroupLeft.GroupLeftResponse.Builder builder = (ProtoGroupLeft.GroupLeftResponse.Builder) message;
         final long roomId = builder.getRoomId();
         final long memberId = builder.getMemberId();
 
         Realm realm = Realm.getDefaultInstance();
-
         realm.executeTransaction(new Realm.Transaction() {
             @Override
             public void execute(Realm realm) {
 
-                RealmRoom realmRoom =
-                        realm.where(RealmRoom.class).equalTo(RealmRoomFields.ID, roomId).findFirst();
+                RealmRoom realmRoom = realm.where(RealmRoom.class).equalTo(RealmRoomFields.ID, roomId).findFirst();
                 if (realmRoom != null) {
                     realmRoom.deleteFromRealm();
-
-                    G.onGroupLeft.onGroupLeft(roomId, memberId);
                 }
 
-                RealmRoomMessage realmRoomMessage = realm.where(RealmRoomMessage.class)
-                        .equalTo(RealmRoomMessageFields.ROOM_ID, roomId)
-                        .findFirst();
-                if (realmRoomMessage != null) {
-                    realmRoomMessage.deleteFromRealm();
-                }
+                if (!builder.getResponse().getId().isEmpty()) { // if own send request for left
+                    RealmRoomMessage realmRoomMessage = realm.where(RealmRoomMessage.class).equalTo(RealmRoomMessageFields.ROOM_ID, roomId).findFirst();
+                    if (realmRoomMessage != null) {
+                        realmRoomMessage.deleteFromRealm();
+                    }
 
-                RealmClientCondition realmClientCondition = realm.where(RealmClientCondition.class)
-                        .equalTo(RealmClientConditionFields.ROOM_ID, roomId)
-                        .findFirst();
-                if (realmClientCondition != null) {
-                    realmClientCondition.deleteFromRealm();
+                    RealmClientCondition realmClientCondition = realm.where(RealmClientCondition.class).equalTo(RealmClientConditionFields.ROOM_ID, roomId).findFirst();
+                    if (realmClientCondition != null) {
+                        realmClientCondition.deleteFromRealm();
+                    }
+
+                    if (G.onGroupLeft != null) {
+                        G.onGroupLeft.onGroupLeft(roomId, memberId);
+                    }
                 }
             }
         });
-
         realm.close();
+
+
     }
 
     @Override
